requirements.txt
requests
python-dotenv
tqdm
openpyxl
streamlit
pandas